图片地图绘制

+++ src/core/EntityPackage/Entity.java
@@ -3,7 +3,7 @@ package core.EntityPackage;
 import java.awt.image.BufferedImage;
 
 public class Entity {
-    public double x , y;
+    public int x , y;
     public double speed = 4.0;
 
     public BufferedImage up1, up2, up3, up4, down1, down2, down3, down4, left1, left2, left3, left4, right1, right2, right3, right4;
